Refactor 7
