+++ govstack_api/building_blocks/bb_digital_registries/urls.py
@@ -8,6 +8,7 @@ from govstack_api.building_blocks.bb_digital_registries.views import SearchRecor
 from govstack_api.building_blocks.bb_digital_registries.views import SingleRecordAPI
 from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView, SpectacularRedocView
 
+from govstack_api.building_blocks.bb_digital_registries.views.single_field_view import SingleFieldView
 
 urlpatterns = [
     path('data/<str:registryname>/<str:versionnumber>/create', SingleRecordAPI.as_view(), name="create_registry_record"),
@@ -17,7 +18,7 @@ urlpatterns = [
     path('data/<str:registryname>/<str:versionnumber>/update-or-create', UpdateOrCreateRecordView.as_view(), name="update_or_create_registry_record"),
     path('data/<str:registryname>/<str:versionnumber>/<str:ID>/delete', SingleRecordAPI.as_view(), name="delete_registry_record"),
     path('data/<str:registryname>/<str:versionnumber>/<str:uuid>/read-value/<str:field>.<ext>',
-         SingleRecordAPI.as_view(), name="read_field_from_registry_record"),
+         SingleFieldView.as_view(), name="read_field_from_registry_record"),
     path('data/<str:registryname>/<str:versionnumber>', MultipleRecordAPI.as_view(), name="list_registry_records"),
     path('data/<str:registryname>/<str:versionnumber>/update-entries', MultipleRecordAPI.as_view(), name="update_registry_records"),
     path('data/myPersonalDataUsage/1.0', PersonalDataAPI.as_view(), name="get_personal_data"),
